+++ xonsh/parsers/base.py
@@ -1352,13 +1352,6 @@ class BaseParser(object):
                                    col_offset=p1.lexpos)
         p[0] = p0
 
-    def p_async_stmt(self, p):
-        """async_stmt : async_funcdef
-                      | async_with_stmt
-                      | async_for_stmt
-        """
-        p[0] = p[1]
-
     def p_suite(self, p):
         """suite : simple_stmt
                  | NEWLINE INDENT stmt_list DEDENT
@@ -1734,57 +1727,6 @@ class BaseParser(object):
             rtn = [x]
         return rtn
 
-    @docstring_by_version(
-        v34="""atom_expr : atom trailer_list_opt""",
-        v35=\
-        """atom_expr : atom trailer_list_opt
-                     | await_tok atom trailer_list_opt
-        """
-        )
-    def p_atom_expr(self, p):
-        lenp = len(p)
-        if lenp == 3:
-            leader, trailers = p[1], p[2]
-        elif lenp == 4:
-            leader, trailers = p[2], p[3]
-        else:
-            assert False
-        p0 = leader
-        if trailers is None:
-            trailers = []
-        for trailer in trailers:
-            if isinstance(trailer, (ast.Index, ast.Slice)):
-                p0 = ast.Subscript(value=leader,
-                                   slice=trailer,
-                                   ctx=ast.Load(),
-                                   lineno=leader.lineno,
-                                   col_offset=leader.col_offset)
-            elif isinstance(trailer, Mapping):
-                p0 = ast.Call(func=leader,
-                              lineno=leader.lineno,
-                              col_offset=leader.col_offset, **trailer)
-            elif isinstance(trailer, str):
-                if trailer == '?':
-                    p0 = xonsh_help(leader, lineno=leader.lineno, col=leader.col_offset)
-                elif trailer == '??':
-                    p0 = xonsh_superhelp(leader,
-                                         lineno=leader.lineno,
-                                         col=leader.col_offset)
-                else:
-                    p0 = ast.Attribute(value=leader,
-                                       attr=trailer,
-                                       ctx=ast.Load(),
-                                       lineno=leader.lineno,
-                                       col_offset=leader.col_offset)
-            else:
-                assert False
-            leader = p0
-        if lenp == 4:
-            p1 = p[1]
-            p0 = ast.Await(value=p0, ctx=ast.Load(), lineno=p1.lineno,
-                           col_offset=p1.lexpos)
-        p[0] = p0
-
     def p_atom(self, p):
         """atom : lparen_tok yield_expr_or_testlist_comp_opt RPAREN
                 | lbracket_tok testlist_comp_opt RBRACKET
@@ -2070,23 +2012,6 @@ class BaseParser(object):
                            (hasattr(p1, '_real_tuple') and p1._real_tuple):
             p1.lineno, p1.col_offset = lopen_loc(p1.elts[0])
         p[0] = p1
-
-    @docstring_by_version(
-        v34="""item : test COLON test""",
-        v35=\
-        """item : test COLON test
-                | POW expr
-        """,
-        )
-    def p_item(self, p):
-        lenp = len(p)
-        if lenp == 4:
-            p0 = [p[1], p[3]]
-        elif lenp == 3:
-            p0 = [None, p[2]]
-        else:
-            assert False
-        p[0] = p0
 
     def p_comma_item(self, p):
         """comma_item : COMMA item"""
